web_app_v2/backend/requirements.txt
# Backend dependencies for FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
PyYAML>=6.0
openai>=1.0.0

# For session storage (optional - can use memory instead)
redis>=5.0.0

# For WebSocket support
websockets>=12.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
